Hash derivations based on their realized inputs
As it turns out, this simplifies the process even more
because we can find realizations as we walk up the tree.

I added hooks for using realizations that aren't in the store yet,
which is useful once substituters are in play,
but aren't necessary for the introducing the early cutoff optimization.

Fixes #38