Add option to drop privileges for builders when running server as root
The `zb serve --build-users-group` option
functions largely the same as in Nix.
If we're running as root, we default to `--build-users-group=zbld`.

This required a handful of tweaks
to chmod/chown build and store directories as appropriate.

Annoyingly, there isn't a standard library function
to list the Unix users in a group
(because the operation is somewhat ill-defined, as I discovered).
I implemented one that invokes `getent` and falls back to `/etc/group`.
I think this should be good enough for most cases I can think of.

Fixes #51