Add `luacode.Value.IdenticalTo` method
This makes equality vs. identity checks more explicit
and handles a few weird edge cases with them.
We no longer use `unique.Handle`
to build up the constant table, but it's okay
because we can use a hashing scheme
to reduce the number of comparisons per constant table insert.
This is roughly equivalent to what was happening before,
but with less runtime machinery.