Create new Lua interpreter per frontend eval call
This prepares the frontend for a more isolated import system.

Added a benchmark to measure performance.
Originally was ~80μs,
but about half was spent in parsing the prelude.
Added a `go generate` step that precompiles the prelude,
and got it down to ~40μs.
There's certainly more tuning that can be done,
but this is more than good enough to start.

Updates #75