Attempt to unmount files while cleaning up chroot sandbox
Created a variant of `os.RemoveAll`
that calls `umount(2)` on each file before removing it.
If it fails with an error that indicates it was a mountpoint,
then it skips recursion.
This should make it safe to call on the chroot,
even if there are previously failed unmount calls,
and it will try to clean up as much as it safely can.

Fixes #70