New build resource RPCs
The peer-to-peer architecture of the backend RPC layer
makes remote builds and a possible HTTP transport layer difficult.
Furthermore, the streaming log aspect means that a slow receiver would actually slow down builds.

To solve this, I reworked the RPC layer to be stateless.
Starting a realization creates a new "build" resource
which can be polled to get the latest status.
A new "read log" RPC is used to read the builder logs,
which are unconditionally persisted to the store database.
An age-based periodic garbage collection cleans up old logs,
but they are by default retained for a week.

This had some downstream effects on the frontend and the CLI code:

- Multiple targets can now be specified in one build over JSON-RPC.
  The CLI has been updated accordingly, and this fixes #22.
- I realized that the logging change would need to be plumbed through the frontend in some way.
  However, it didn't seem right that the frontend API should change,
  so I introduced a `frontend.Store` interface that defines what operations the frontend needs,
  and provide appropriate implementations for the CLI and in test.

Finally, the new code in realize.go is more disciplined about connection management,
so it should form as a better basis for #32.

Fixes #97
Fixes #22