Expand acceptable `system` attributes
Now using roughly the same parsing logic as in LLVM,
so triples can be used as-is.
Created a fuller test suite that handles a lot of edge cases.
Existing tuples (`x86_64-linux`, `aarch64-linux`, `aarch64-darwin`, and `x86_64-windows`)
have the same semantic meaning and will continue to be recognized.

Specification document needs to be updated to match.

Updates #110