Add `bytebuffer.Creator` interface
In preparation for Mach-O multi-pass analyses (#134),
I want to give the users of the zbstore package control over resource usage.
As part of that work,
I developed an interface for creating temporary files.

There are a couple existing places that create temporary files,
so I reworked them to use `bytebuffer.Creator`.
This pushes the naming policy up to cmd/zb
and enables tests to use in-memory buffers,
speeding up test execution.

I promoted the `internal/bytewriter` package
up into the new `bytebuffer` package.
The `bytebuffer.Creator` needs to be exported
in order for zbstore to use it in parameters.
It felt clumsy to include the interface in zbstore,
so having a separate generic package felt the most natural.