Rewrite UUIDs on certain Mach-O binaries with self-references
To reflect the change in executable content,
the UUIDs are regenerated with a UUID based on the contents.
Changing the UUID also changes the code signature,
so rewrites are now returned in analysis in order of execution.

This required doing multiple passes on Mach-O files,
since we don't know whether to zero out the UUID
until we encounter the code signature at the end of the file.
Thankfully, we only have to buffer in at most one architecture image at a time,
so the amount of disk required should be fairly minimal.
If a store object does not contain a Mach-O file,
then content addressing does not require additional storage.

The internals of the content addressing code have been heavily restructured,
but the API is fundamentally the same.
While testing this code, I was finding it difficult to debug failures
purely based on hashes not matching.
I factored out the writing process as an unexported function
to obtain the cleartext in the test,
but kept the assertions against `SourceSHA256ContentAddress` as before.
This makes it significantly easier to debug failing test cases
while still providing the same confidence in the exported API.

The content addressing logging is now exposed as a generic callback option.
The zb CLI can now properly plumb context to the log calls.

Fixes #134